# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Zenstats — 自托管网站分析服务，提供隐私友好的 Web 分析 API。\n\n## 认证方式\n- **BearerAuth**: 在 Authorization 头中传入 `Bearer \u003cJWT Token\u003e`，用于管理类 API\n- **APIKeyAuth**: 在 Authorization 头中传入 API Key，用于统计查询 API\n\n## 统一响应格式\n成功: `{ \"code\": 200, \"message\": \"success\", \"data\": ... }`\n失败: `{ \"code\": \u003cstatus\u003e, \"message\": \"...\", \"error\": \"...\" }`\n\n## API 分组\n- **认证** — 登录、注册、令牌刷新、密码重置、邮箱验证、系统初始化\n- **站点管理** — 站点 CRUD、域名验证、屏蔽规则（IP/域名/国家）\n- **统计分析** — 聚合指标、时间序列、维度细分、实时访客、CSV 导出\n- **漏斗分析** — 漏斗 CRUD + 转化率分析\n- **目标管理** — 转化目标 CRUD\n- **事件采集** — 前端埋点 SDK 上报端点\n- **数据导入** — GA4 历史数据导入与查询\n- **用户** — 自定义搜索引擎、子账号管理、配额查询、个人资料\n- **API Key** — API Key 创建、列表、删除\n- **管理员** — 用户/站点/套餐管理、系统配置、系统统计\n- **健康检查** — 服务及依赖连通性检查",
        "title": "Zenstats API",
        "termsOfService": "https://github.com/zenstats/zenstats",
        "contact": {
            "name": "Zenstats",
            "url": "https://github.com/zenstats/zenstats",
            "email": "wrpota@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/user/sub-accounts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前用户的子账号列表。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取子账号列表",
                "responses": {
                    "200": {
                        "description": "子账号列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.SubAccountListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "为当前用户创建子账号。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "创建子账号",
                "parameters": [
                    {
                        "description": "子账号信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateSubAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.SubAccount"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "邮箱已存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/sub-accounts/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新指定的子账号信息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "更新子账号",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "子账号ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "子账号信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateSubAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "子账号不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除指定的子账号。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "删除子账号",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "子账号ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "子账号不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/sub-accounts/{id}/reset-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "重置指定子账号的密码。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "重置子账号密码",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "子账号ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "新密码",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ResetSubAccountPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "重置成功",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "子账号不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "response.ErrorResponse": {
            "description": "接口返回的错误响应信息",
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer"
                },
                "error": {
                    "description": "错误详情",
                    "type": "string"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        },
        "response.SuccessResponse": {
            "description": "接口返回的成功响应信息",
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应信息",
                    "type": "string"
                }
            }
        },
...
        "types.CreateSubAccountRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 8
                }
            }
        },
...
        "types.ResetSubAccountPasswordRequest": {
            "type": "object",
            "required": [
                "new_password"
            ],
            "properties": {
                "new_password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 8
                }
            }
        },
...
        "types.SubAccount": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_seen": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "types.SubAccountListResponse": {
            "type": "object",
            "properties": {
                "current_count": {
                    "type": "integer"
                },
                "has_permission": {
                    "type": "boolean"
                },
                "max_sub_accounts": {
                    "type": "integer"
                },
                "sub_accounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.SubAccount"
                    }
                }
            }
        },
...
        "types.UpdateSubAccountRequest": {
            "type": "object",
            "required": [
                "name",
                "status"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "active",
                        "suspended"
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "description": "在 Authorization 头中传入 API Key（无需 Bearer 前缀）。在站点管理页面创建 API Key。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BearerAuth": {
            "description": "在 Authorization 头中传入 `Bearer \u003cJWT Token\u003e`。获取方式：调用 /auth/login 或 /auth/register。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}